refactor: refactored into a single component file
